vulkan: add packed 16-bit formats
This can be used to map P010 etc. natively, without requiring conversion
or up/down-shifting. More importantly, these are needed to support
packed 16-bit planar formats.

Also add the corresponding DRM fourcc, for the one format with a
matching code.

It's worth pointing out that this implementation doesn't add any special
`pl_fmt` signalling for the 10-bit-in-16-bits padding nonsense - it
literally just sets the precision to 10 bits while the host size is 16
bits. This reflects the fact that, according to the host, you are
uploading 16-bit words - the lowest bits just get cut off due to the
lack of texture precision.

This doesn't really matter in practice because the most important
consumers (`pl_upload_plane` et al.) don't end up using this format
anyways, they usually pick the normal `r16` format for padding r10x6
input planes. (This is also important for compatibility with OpenGL
etc.)

We guard the vulkan version to pull in the un-suffixed names.