utils/upload: support uploading non-native endian data
In practice, this refers to big endian formats. This is done using the
new `pl_buf_copy_swap` helper. Note that, in theory, it would be doable
to mix this into the `pl_tex_upload_texel` helper to save one memcpy
(from the buf to the tex), but the code overhead for this is simply not
worth it.

We could also save one memcpy by doing a host ptr import (if we support
storable host pointers), but this is also a special case I don't really
care about too much to justify the code cost.

In practice, the only relevant big endian format is PNG, so supporting
it is more of a courtesy than a real consideration.