utils/libav: add `pl_map_hdr_metadata`
This not only exposes the previous logic to users, but also adds support
for AVDynamicHDRPlus. Use this function internally in place of the
previous duplicated code, and also get rid of some sanity masking logic
that was of dubious importance (especially now that `nominal_min/max`
are decoupled from the HDR metadata).

For HDR10, we only map the fields relevant/interesting for us, ignore
the rest. (Which is also unused in practice)