tone_mapping: default to ST2094 when metadata present
This generally produces good results in such a case. ST2094-40 is better
than ST2094-10 even in the absence of metadata, because the relevant
parts of ST2094-10 (the average->average mapping) are also implemented
in the ST2094-40 fallback path, and more importantly, the latter also
implements proper black point adaptation (to avoid black crush).

In the future, ST2094-40 (possibly with some tweaks) might become the
one-size-fits-all default tone-mapping method for all scenarios.