utils/libav: don't crash on rgb565be or x2rgb10be
These two formats currently result in an assertion failure due to being
encoded in a format that implicitly assumes 32-bit word aligned
sampling. Worryingly, there's no way to tell (in libavutil), whether or
not we're dealing with such a format or a "normal" big endian format.

So simply detect this case by looking for "shifted" components.