colorspace: refactor HDR levels fields
So, I was growing increasingly dissatisfied with the need to override
the HDR mastering display metadata for PQ inputs, and wanted to cleanly
separate concerns here.

After this refactor, the *mastering metadata* remains purely
informative, and all of the relevant defaulting logic goes on inside
`pl_color_space`. This field will also be consulted for any operations
relevant to tone-mapping, and can take into account both per-scene
metadata as well as static metadata.

As an aside, I decided to just completely drop back-compatibility with
the deprecated `sig_*` fields, because it's been years.