tone_mapping: add SMPTE ST2094-10 and ST2094-40 Annex B
The implementation for ST2094-10 is straightforward by just copying
equations from the corresponding SMPTE document. While this relies on
metadata for the best results, we can get good enough results by just
taking the avg to be 10.

For ST2094-40, includes modifications from the ATSC A/341 amendment. I
had to omit the section about fixing the first bezier point (P[1]),
because it resulted in a completely broken output for some samples (due
to extreme slopes in the knee function). Also includes black point
adaptation, based on specification BT.1886 - something that was notably
missing in the original SMPTE spec.

Omits the parts of the spec we can't faithfully implement, such as the
detail processing BS, support for overlapping elliptical windows, and
the ST2094-10 subjective gain/offset controls missing in HDR10+
metadata.

Closes: https://code.videolan.org/videolan/libplacebo/-/issues/244
Closes: https://code.videolan.org/videolan/libplacebo/-/issues/245