colorspace: pass scene_max through luma coeffs
Instead of just blindly taking the PL_MAX3() of this array. This should
serve to hopefully fix a disconnect between what we consider to be the
relevant metric (max luminance) and what HDR10+ thinks is important
(maximum component).

Arguably, doing it this way is worse for PL_TONE_MAP_RGB, but since we
default to the BT.2446a-style PL_TONE_MAP_LUMA methods, this should give
us better results in practice.