vulkan/utils_gen: fix build on vulkan-headers 1.3.241
This version of the headers adds support for the vulkansc API, which
pulls in a bunch of definitions that we don't care about.

To properly exclude these, we need to start blacklisting all types not
defined in the vulkan environment. Also leverage this to start
blacklisting platform-specific crap in a more principled way.

This is probably still not the correct way to do it, but it works for
now.

Fixes: https://code.videolan.org/videolan/libplacebo/-/issues/249
Fixes: https://github.com/haasn/libplacebo/issues/155