renderer: disable tone-mapping params for overlays
The default BPC is good enough for overlays, no need to generate a full
LUT here. Inverse tone-mapping of overlays is probably never intended,
anyway. This will definitely break some sort of HDR anime softsubbed
signs, but not only is the subbing scene dead, this would require HDR
subtitle formats to start existing. (And for the love of god, somebody
just move softsubbed signs to an actual video track already!)

Fixes: https://github.com/mpv-player/mpv/issues/11345