demos/plplay: major refactor of HDR override section
Instead of these complicated bools and structs that get updated from
different places at diferent times, simplify this set of fields into
something easier to understand: `force_*` fields which simply
hard-override their respective values if present.

Rather than updating from multiple places, we thread the configured
target swapchain directly into the `update_settings` function so we can
use it directly as the basis for our configured HDR levels, including
displaying the actually-signalled values *as-is* in the absence of an
override.

We also allow the ICC luma hack to be configurable, forcing
icc_params.max_luma to always be either the detected luminance or the
configured luminance (falling back to the default SDR luminance).