pl_string: add pl_str_append_raw
To avoid overhead of adding a bunch of redundant \0 terminators in cases
when we want to frequently append small strings, such as inside
pl_str_append_asprintf_c.

Also changes pl_str_append to always omit the \0 even when appending an
empty string, so it can effectively be used as a way to force strings to
be \0-terminated at the end of such a function.