dispatch: only execute string templates when compiling
When re-using existing shaders, we now ignore the string builders, and
defer execution to the case where the shader needs to be rebuilt.

This commit needs a lot of special care to ensure that we don't conflate
pass *signatures* (which are now based on the output of
pl_str_builder_hash, unstable values that aren't expected to persist
across program invocations) and the hash used for shader cache entries,
which require some special finesse to ensure their values don't change
as a result of this change.

I've tested this commit to make sure that cached_program entries receive
the exact same hash as before.