shaders/colorspace: rewrite peak detection shader
As a result of the significant refactors to the peak detection logic,
it's now possible for us to move most of the complicated peak detection
state update function to the CPU as well.

This allows us to significantly reduce complexity, and also greatly
expand the functionality of peak detection. It also greatly improves
accuracy, enough so that we no longer even need to simulate the rounding
errors in the test code.

This commit merely rewrites the algorithm, but does not yet add new APIs
or deprecate existing ones.

Closes: https://code.videolan.org/videolan/libplacebo/-/issues/256