shaders/colorspace: use linear stretch only if requested
Rather than always hard-coding this for BPC. The reasoning here is that
many tone-mapping functions end up doing black point adaptation in a way
that is subjectively quite different from a linear stretch. So we should
always prefer the requested tone-mapper, even for pure BPC operations.

This fixes the remaining discontinuity when dst_max > src_max, and
allows us to drop the "safety clamp" in the peak detection branch.

Fixes: https://code.videolan.org/videolan/libplacebo/-/issues/254