shaders/colorspace: remove pl_peak_detect_params.overshoot
During the rewrite of the peak detection shader, we unintentionally
switched to doing averaging, overshoot, scene change detection etc. in
PQ space, rather than in linear nits. This essentially broke the
overshoot from functioning in the intended way.

Additionally, as a result of the previous changes, overshooting has less
of a clear benefit than before. It also causes issues due to not being
retroactively applied if the params suddenly change.

As a result of all these considerations, simply remove it.

Fixes: daed681e04e7ccd36b1302940ec2c2001b9b67ee