shaders/colorspace: fix scene change hysteresis
The rewrite to peak detection made this logic happen on PQ, when it was
clearly intended to be calculated with raw nits values (hence the log
calls). Instead of using abs(log(new / old)), switch to abs(new - old)
in PQ space, which is already quasi-logarithmic.

Use the experimentally determined parameter 0.25f to convert from dB
gain to PQ delta, to roughly preserve backwards compatibility with the
option's documentation.

Fixes: daed681e04e7ccd36b1302940ec2c2001b9b67ee
Fixes: https://github.com/mpv-player/mpv/issues/11406