shaders/colorspace: add pl_color_map_params.visualize_rect
Affords users some control over where to play the tone-mapping
visualization. This is not an exact science since it is relative to the
effective rendered crop and scaled in units of 0-1, but it's surely
better than nothing.

Also tweak the display parameters slightly to more visually distingish
the graph from surrounding regions.