renderer: fix alpha assertions
After the change in alpha handling as a result of splitting up
pass_output_target, we broke a few hard-coded assumptions about the
alpha representation. Make these more robust.

Also, while inspecting this code, I realize that the last pass is always
guaranteed to output a shader, so we can drop the blit path for the time
being.

Fixes: https://github.com/haasn/libplacebo/issues/161