shaders/colorspace: add frame histogram measurement
We only use this to refine the measured peak value, to filter out
outlier pixels that are much brighter than the rest of the scene. The
default of 99.995% is very conservative, and will not fix extreme
examples such as when there's a very bright logo on an HDR demo reel,
but more transparently enhances real content.

Closes: https://code.videolan.org/videolan/libplacebo/-/issues/236