shaders/colorspace: fix scene change hysteresis (again)
This factor was accidentally off by an order of magnitude. It was also
too high, as a result of being estimated only for ratio 2 instead of
higher coefficients.

To be clear, this coefficient is supposed to be scaled such that `3 *
log10_pq` represents the pq delta that will result in a doubled signal
level, which is roughly 0.04 for low brightness levels. This turns out
to be almost 0.01 PQ, so we can remove the confusing (and misleading) dB
comment and just (correctly) document it as being scaled to 1% PQ steps.

Fixes: 6b7e84c0afa19ec6efa4988ad9e2b2b8b385f2a9
Fixes: https://github.com/mpv-player/mpv/issues/11406