shaders/colorspace: remove hard-clamp
Clamping the input *RGB* values to the input *luminance* is an obvious
no-no. I'm not sure what originally motivated this clamp, but I can no
longer find a sample where it's required - most likely as a result of
our current default tone-mapper doing a good job on extreme highlights.

I think this was a hack, and it's a stale one.