shaders: parametrize `ident_t` printf fmt and initializer
I want to switch from using string-based identifiers to using integers.
To make this transition possible in as pain-free a manner as possible,
we first parametrize the current printf formatter / initializers.

Using $ as an identifier is highly dubious, and some might consider
immoral, but it's accepted by our current compilation target (GCC and
clang), is short, extremely unlikely to clash with other defines from
other headers, and makes the syntax visually similar to the $foo-style
pragma macros that will replace it.