shaders: refactor shader identifier naming
Preliminary work for an upcoming refactor which will treat identifiers
directly as unsigned shorts.

Although space in a 16-bit integer is somewhat tight, collisions should
hopefully be impossible in practice because shader namespaces can only
be 8-bit in theory (and in practice, are usually below 8), and even 256
descriptors per shader is already unlikely - let alone the ~16k that
would be required to hit a collision.