format: cast to ushort before calling print_hex
This may seem counterintuitive because `print_hex` is implemented for
uint anyway, but this lets the compiler know that upper bits are zeroed,
as we know for sure, because it is `%hs`. This in turn allows compiler
to prune always unused switch cases. `print_hex` is static and called
only with ushort, so even if not inlined this DCE will happen.