shaders: pool small temporary allocations
These are used for things like pl_memdup() inside pl_var() et al.

Ideally, I would like to have a static scratch buffer pre-allocated
inside the `pl_shader_t` itself, but this is not possible as a
consequence of the sh_subpass shader merging design - after the subpass
is merged, pointers to it will point to undefined memory (unless we
reparent the entire subshader).

For the same reason, we can't gracefully handle shader subpass merging
without having to steal the scratch buffer and reparent it onto the main
shader also, effectively wasting it. In the future I may revisit
`sh_subpass` to solve both problems.

This commit slightly reduces the number of allocations required to
generate shaders.