colorspace: add IPTPQc4 model definitions
Will be used for the upcoming major refactor of the tone mapping / gamut
mapping pipeline.

We deliberately deviate from the existing state of the art (ICtCp,
IPTPQc2, etc.). We use the Ebner & Fairchild 1998 IPT matrix instead of
the modified ICtCp matrix, because the former has better hue linearity -
and we care about using this for gamut mapping applications. We deviate
from the Dolby-style IPTPQc2 and continue to use a 4% crosstalk, as this
is the smallest crosstalk parameter that prevents concavities inside the
IPT color space. We want to avoid concativities in order to use this for
hue-preserving gamut mapping.