vulkan/context: always initialize all queues
To improve compatibility with other users of this VkDevice (notably
FFmpeg), we need to create e.g. decoding/encoding queues, when
available.

In practice, these shouldn't show up unless the necessary extensions are
enabled, so this is a no-op in most circumstances. At worst, we waste a
tiny bit of driver memory and overhead.

To preserve API compatibility, we still only announce the basic queues
in `pl_vulkan.queues`. Deprecate the field on account of this
limitation.