vulkan/tex: pass correct DRM format modifier to fmt check
This call was currently hard-coded to pass shared_mem.drm_format_mod,
which is undefined behavior on DMABUF export (not import). Fix it by
instead choosing (arbitrarily) the first modifier from the list of
supported modifiers. The spec says we should check *all* modifiers, but
in practice this call only exists to catch bugs in the other fields, so
checking only one should suffice.