vulkan/utils: add feature normalization helper
Vulkan decided to handle promoted features in a very stupid and
inconsistent way, where there are two different ways to specify promoted
features and the application must use consistently either one or the
other. Additionally, some promoted features do *not* have corresponding
extension structs. So the only way to handle this all in a sane and
consistent way is to write a helper function to properly normalize a
features chain.