d3d11: remove static texels count limit
Direct3D 11 runtime only guarantees that allocations within the maximum
resource size are supported, but an app can try to allocate more memory
for a resource than the maximum resource size specifies. That is, the
Direct3D 11 runtime allows these memory allocation attempts in the event
that the hardware might support them. Instead hardcoding limit here,
allow it to be resolved during runtime.

Fixes: https://github.com/mpv-player/mpv/issues/11663