vulkan: switch (un)lock_queue to uint32_t
For consistency with both the Vulkan spec and the revised version of the
lavu vulkan patchset. This was only chosen as `int` for compatibility
with previous versions of that patchset.

This API has no users other than an experimental mpv branch, so it
should be inconsequential to break. In either case, machine
interpretation of non-negative int and uint32_t should be identical on
all sane platforms.