gamut_mapping: replace PQ EOTF by LUT
With linear interpolation, this gives enough precision even on a
256-element LUT, to pass the tests.

1232.685 ms -> 422.975 ms on my machine, almost 3x as fast.

It's worth pointing out that the inverse function, pq_oetf, is not a
major bottleneck and therefore does not need to be optimized. This is
quite fortunate, because quantizing the optical domain of PQ would be
rather terrible for precision.