shaders/colorspace: add HDR contrast recovery
After a few iterations I settled on this, which is essentially a
HDR-specific sharpening filter, that tries to maintain perceptual local
contrast differences (in the HDR source) while tone-mapping global
contrast.

The basic idea is to split the image into HF and LF components and apply
the tone-mapping algorithm only to the LF component. Set to fairly
conservative values by default.

Closes: https://code.videolan.org/videolan/libplacebo/-/issues/261