renderer: automatically expand/constrain distortion canvas
Instead of implicitly clamping the output image to the originally
specified dst_rect, this allows the result of applying a distortion
matrix to effectively exceed the indicated target crop area.

Of course, this is probably surprising to some users, but then again,
distortion isn't exactly a common use case - and I would argue that this
makes arbitrary rotation work much more intuitively.