shaders/colorspace: adjust 3DLUT resolution
Hue is the most important axis, and even 45 is not enough by far.
Increase it considerably. Ideally, we'd use several hundred here, but
this results in too-slow generation. I picked the values by careful
experimentation, and decided that a power-of-two width/height are
perfectly sufficient for our use case, only the luma axis really wants
an odd value (to contain 50% grey).

Also slightly increase chromaticity resolution, and decrease I precision
(which is not that important in practice) to compensate.