gamut_mapping: multi-thread 3DLUT generation
Absolutely massive speedup on typical systems, for very little overhead.

Probably not the smartest way to write this loop (should take into
account number of CPUs) but I don't have conveniently accessible
helper functions for that.

Incidentally, this absolutely butchers the CPU time logging, and we need
to stop using clock() to fix that.