gamut_mapping: soft-clip in RGB after mapping
The reason this new curve works so well is because it incorporates a bit
of hard clipping in RGB space, which is functionally equivalent to
applying a sort of very soft intensity-dependent rolloff curve. Make
this clipping soft rather than hard, and also an explicit part of the
gamut mapping function - so it shows up on the graphs, and also to make
the function well-behaved.

This does distort hues very slightly, but hopefully in an aesthetically
pleasing way.