tools/glsl_preproc: properly ensure uniqueness of names
If expressions would end up with the same slugify'd name, we need to
properly deduplicate them, even in the variable references. Do this by
directly mutating the variable object. Store the object itself in the
list of refs, instead of the pre-formatted string, and override the
__str__() method to format it as-needed.

Note that this will probably still break if we hit the 'reuse existing
entries' branch of the add_var_raw() method. Oh well.