shaders/sampling: enable antiringing for EWA scalers
Based on two-pass design where the main interpolation kernel is combined
with a smaller, second gaussian (non-ringing) filter, with high exponent
(hard-coded as 8 for performance/quality reasons) to bias it towards
low/high values respectively.

This is unfortunately rather slow (about 50% slowdown), but much, much
higher quality than all other approaches I came across. It may be
possible to optimize it slightly more, (or just drop the numbers to make
it faster at the cost of quality), but for now, this represents a vast
quality improvement over the status quo of no antiringing.