demos/plplay: measure and report CPU timings
I first wanted to make a nice graph here like with the shader stats but
eventually bikeshedding got the best of me and I decided to just delete
it because I couldn't figure out a clean way to integrate it into the
nuklear interface.

Besides, not needing to draw a graph vastly simplifies this already
overly-verbose debugging / logging code.