tests/gpu_tests: fix peak detection source data bug
Somehow a deviation in the previous (deband) test resulted in this test
pattern being subtly different from the version actually being tested,
and only on certain platforms (NVIDIA GL, but not AMD GL). While that
is certainly odd in its own right, this commit fixes a straightforward
logic bug in the test.

Fixes: https://code.videolan.org/videolan/libplacebo/-/issues/252#note_401483