renderer: always set HDR levels from ICC profile
Even if `acquire_image` is false. Otherwise, this will not happen on
re-draws with changed image state.

Not entirely sure if this is kosher, but it seems to work fine in
practice for me. Maybe the more principled solution would be to move the
acquision of the ref image to the first pass_init().