options: add new options parsing API
After much deliberation and API revision, this is the form I settled on.
The new `pl_options` basically serves as a helper to refer to an entire
set of params structs at once, plus the ability to treat them in a
manner reminiscient of a key/value dict.

In addition to this, we add a massive internal options list, mapping
(nearly) every single option known to libplacebo. This will serve as an
authoritative and up-to-date way for downstream projects to pipe
arbitrary arguments into libplacebo, as well as make it easier to e.g.
save and load pl_render_params across ABI boundaries. (We have this use
case in e.g. VLC)

One of the main ugly facts about the current options implementation is
that we a tendency to produce grossly overinflated representations of
floating point numbers, but this is mostly a consequence of using
pl_str_append_asprintf_c and its corresponding low-quality float
printing implementation.