filters: add pl_filter_functions/configs
To replace the old remnants of the pl_filter_preset /
pl_filter_function_preset structs that were built around the v1
configuration system.

This brings the treatment of filter functions and configs in libplacebo
to be in-line with the newer API conventions established by
pl_gamut_map_function, pl_tone_map_function etc.

I decided not to PL_DEPRECATE the old configuration API yet, because
porting all software to it will be a nightmare, and deprecating it will
also break pl_renderer (which still relies on this struct for now). So
I'll leave it in limbo until I can decide what to do.