filters: add pl_filter_config usage semantics
This will allow us to mix both types of filters in a single filter list,
while preserving the ability of the API user to know which type of
domain a filter is legal to be used in.

Additionally, add a 'recommended' field to allow trimming down the list
of filters to a shorter list containing only interesting values.