shaders/sampling: redesign polar anti-ringing shader
I have improved the AR kernel by combining a number of techniques:

1. Instead of raising the color value to a large exponent (and then
   taking the inverse exponent later), we can get away with simply
   multiplying the large exponent directly into the weight. This is much
   better for numerical accuracy as it means the error is accounted for
   in wsum, while the original color is left as-is.

2. Instead of taking pow(color, k), take pow(color+1, k). This keeps
   numbers within the sane floating point range without substantially
   altering the result of the powermean calculation. (Thanks to @bjin
   for this trick!)

3. To prevent issues with all-zero wsum, put a small lower bound on the
   computed weight. (I have no idea how ww can ever be 0 here, but
   apparently it can...)

4. Finally, to handle edge case where hi < lo, replace the color value
   by the average of the two in this case.

The combination of the above techniques allows us to raise the kernel
exponent to a significantly larger number (k=32 works well, any higher
introduces more artefacts due to being too close to a true max() - we
want an exponent that preserves a good balance between the weighting
kernel and the maximum value), substantially improving quality.

The downside is that this technique is also quite a bit lower, as a
result of the more involved weight summation and higher exponent
requiring more multiplications.

See-Also: https://github.com/bjin/mpv-prescalers/issues/57