meson: don't use declare_dependency for UCRT import lib
Starting from meson v1.2.1 it is unallowed to link to uninstalled static
libraries. This is how declare_dependency(link_with: ...) is treated. In
fact we want to use UCRT import lib only internaly and only when linking
shared library and/or executable. We don't want for it to be treated as
dependency of anykind, just as a directive for linker. Similar to linker
script.

Unfortunately there is not way to specify link_depends for all targets,
so we have to specify it for all of them.

Thanks to @xclaesse for his patience and idea with link_depends.

Fixes: https://github.com/haasn/libplacebo/issues/191