shaders/colorspace: allow smoothing period of 0.0
To completely disable smoothing, since even a smoothing period of 1.0
implies a small amount of smoothing.

This does technically regress the status quo behavior when passing 0
(which would previously see it being defaulted to 100), but with the way
options are handled these days (via the PEAK_DETECT_DEFAULTS macros) I'm
hoping to phase out the use of such implicit defaults to begin with.

It also doesn't break any current client (I checked all).