tone_mapping: simplify auto-TM selection
The `spline` function has become good enough to outperform all of the
other tone-mapping functions in all of the listed cases. There is no
more need for this hacky auto-selection logic. Just pick spline in all
cases.

The sole exception is with the availability of HDR10+ dynamic metadata
(OOTF), which still outperforms spline as a result of the OOTF baking in
artistic intent.