shaders/colorspace: significantly reduce scene threshold
Based on experimental testing on real content, I determined the "safe"
value for in-scene PQ fluctuation to be around 1%, with hard scene
changes being as low as 2% and generally above 2.5%-3%. So, set the
upper threshold to 3% to be on the safe side.

Fixes problems with sluggish reactions to scene changes in a number of
problematic scenes.