shaders/colorspace: don't warn on solid black frame
Since 31082eee7a it's possible for frame_wg_count to not get
incremented, in particular, for all black WGs. This may lead to the
scenario where frame_wg_count is 0 for the frame/slice. Avoid false
positive on the allow_delayed check and also handle this case correctly
by introducing explicit `frame_ready` field which is set to 1 once per
compute shader invocation.

Fixes: 31082eee7a6dc800df5bd7db6d115efcaaa2dd8c