cache: add common/core caching mechanism
Motivated by https://code.videolan.org/videolan/libplacebo/-/issues/299

This mechanism should be more pluggable and less opaque than the
previous ad-hoc dispatch cache, and gives the user better control over
exactly when and where stuff is cached (or not).

I will eventually also add an internal file-based cacher to serve as an
example cache_cb, but that is a plan for the future.