d3d11/pass: switch to pl_cache based caching
Trivial port by just replacing cached_program with
pl_cache_lookup/insert.

The cache key is made up of the shader body only, but we keep an
internal signature in the cache header to verify and invalidate the
cache when the shader compiler changes.

Co-authored-by: Niklas Haas <git@haasn.dev>

Rebased on top of revised pl_cache API. This version is a bit more
wasteful than it could optimally be, since we could be re-using the
buffer from the old cache object to store the new data. But I'd rather
avoid the headache right now.