config: attach version to deprecation macros
To allow tagging deprecated fields with the specific version that
deprecated them. The argument is technically just free-form text, so I
decided to put the major version and API version of the deprecation.
This gives users the most clarity, I think.

It will admittedly make rebasing commits that touch the API version
slightly more annoying, but hopefully this is manageable in practice.

Also be a bit more consistent about the placement of deprecation macros.