renderer: don't assert on a failed icc profile
At least for mpv, if we have a playlist of images, fail opening the ICC
of one image, go forward and then back to the image with the ICC
failure, it'll hit this assertion failure. Not so nice. If there's no
ICC information, it just does nothing so display the image as usual.