filters: Add Power of Blackman and Power of Garamond windows
Libplacebo supports 2 free parameters for windows and in my research (https://github.com/garamond13/Finding-the-best-methods-for-image-scaling) more customizable windows have proven to be superior in quality when scaling images. Also compared to Said window and GNW, for these 2 windows we don't need to do any extra modifications to the code in order to implement them.

Reference for Power of Blackman window: https://github.com/garamond13/power-of-blackman

Reference for Power of Garamond window: https://github.com/garamond13/power-of-garamond-window